requirements.txt
aiogram==3.20.0.post0
asyncpg==0.30.0
python-dotenv

